requirements.txt
# local package
-e .

click==8.0.4
datasets==2.8.0
evaluate==0.4.0
Markdown==3.4.1
numpy==1.23.5
python-dotenv==0.21.0
setuptools==65.6.3
transformers==4.25.1

